Safari-609.1.11.1/LayoutTests/fast/selectors/hover-quirks.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/hover-active-quirks-utility.js"></script>
<style id="testStyle">
</style>
</head>
<body>
<div id="target" class="aClass otherClass" webkit="rocks" style="background-color:green; width:200px; height:200px;"></div>
<div id="console"></div>
</div>
</body>
<script>
description('Test the :hover selector when the document is in quirks mode. To test manually, move the cursor over the green rectangle and reload the page.');

if (window.eventSender) {
    var target = document.getElementById('target');
    var x = target.offsetLeft + target.offsetWidth / 2;
    var y = target.offsetTop + target.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    test(':hover');
} else {
    // To test manually, wait a bit to make sure the browers has updated its state, then run the tests.
    setTimeout(test, 250, ':hover');
}

</script>
<script src="../../resources/js-test-post.js"></script>
</html>
